+++ src/lib/components/home-page/banner.tsx
@@ -3,7 +3,7 @@ import { AppDispatch, RootState } from "@lib/state/store";
 import Image from "next/image";
 import React from "react";
 import { useDispatch } from "react-redux";
-import ButtonDefault from "../../common/buttons/buttonDefault";
+import ButtonDefault from "../common/buttons/buttonDefault";
 import useLocalStorage from "@lib/hook/useLocalStorage";
 import { useRouter } from "next/router";
