+++ src/lib/components/home-page/login.tsx
@@ -2,9 +2,9 @@ import React, { useEffect, useState } from "react";
 import Image from "next/image";
 import { verifyLogin } from "@/api/auth.api";
 import PopupDefault from "@lib/components/common/popups/popupWelcome";
-import Form from "../../common/form/defaultForm";
-import Input from "../../common/input/defaultInput";
-import ButtonDefault from "../../common/buttons/buttonDefault";
+import Form from "../common/form/defaultForm";
+import Input from "../common/input/defaultInput";
+import ButtonDefault from "../common/buttons/buttonDefault";
 import { useRouter } from "next/router";
 import { useDispatch } from "react-redux";
 import { AppDispatch } from "@lib/state/store";
