+++ src/lib/components/home-page/welcome.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 import Image from "next/image";
-import ButtonDefault from "../../common/buttons/buttonDefault";
+import ButtonDefault from "../common/buttons/buttonDefault";
 import PopupDefault from "@lib/components/common/popups/popupWelcome";
 
 const Welcome = () => {
