+++ src/lib/components/layout/defaultLayouts.tsx
@@ -1,9 +1,9 @@
 "use client";
 import React, { ReactNode, useEffect } from "react";
-import Header from "./header/DefaultHeader";
-import Footer from "./footer/DefaultFooter";
+import Header from "./header/defaultHeaders";
+import Footer from "./footer/defaultFooters";
 import { usePathname, useRouter } from "next/navigation";
-import AdminLayout from "./AdminLayout";
+import AdminLayout from "./adminLayouts";
 import { useDispatch, useSelector } from "react-redux";
 import { RootState } from "@lib/state/store";
 import { setTurnOffPopup } from "@lib/state/popup/popupSlice";
