+++ src/pages/_app.tsx
@@ -1,9 +1,9 @@
-import Layout from "@lib/components/layout/DefaultLayout";
 import "@/styles/globals.css";
 import "@/styles/loading.css";
 import type { AppProps } from "next/app";
 import { Provider } from "react-redux";
 import { store } from "../../src/lib/state/store";
+import Layout from "@/lib/components/layout/defaultLayouts";
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
