+++ src/pages/landing/index.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from "react";
 import { useSelector } from "react-redux";
 import { RootState } from "@lib/state/store";
-import Login from "@lib/components/home-page/login";
-import Banner from "@lib/components/home-page/banner";
-import HowToPlay from "@lib/components/home-page/how-to-play";
+import Login from "@/lib/components/home-page/login";
+import Banner from "@/lib/components/home-page/banner";
+import HowToPlay from "@/lib/components/home-page/howToPlay";
 
 const LandingPage = ({ data }: any) => {
   const { popup_login } = useSelector((state: RootState) => state.popup);
